+++ compiler-core/src/type_/error.rs
@@ -679,6 +679,11 @@ pub enum Warning {
         layer: Layer,
     },
 
+    SingleCaseClause {
+        location: SrcSpan,
+        hint: EcoString,
+    },
+
     UnreachableCaseClause {
         location: SrcSpan,
     },
@@ -935,6 +940,7 @@ impl Warning {
             | Warning::InefficientEmptyListCheck { location, .. }
             | Warning::TransitiveDependencyImported { location, .. }
             | Warning::DeprecatedItem { location, .. }
+            | Warning::SingleCaseClause { location, .. }
             | Warning::UnreachableCaseClause { location, .. }
             | Warning::CaseMatchOnLiteralCollection { location, .. }
             | Warning::CaseMatchOnLiteralValue { location, .. }